requirements.txt
Flask>=2.0
Flask-SQLAlchemy
python-dotenv
openai
gunicorn
psycopg2-binary

